[refactor] Slice types
Separate into relevants slices and update references
