+++ client/src/Types/index.ts
@@ -1,0 +1,11 @@
+export * from './redux'
+export * from './objects'
+export * from './parameters'
+export * from './componentProps'
+
+// ============ //
+// DEFINITIONS  //
+// ============ //
+
+export type UserId = string
+export type EventId = string
