+++ client/src/Types/parameters.ts
@@ -1,0 +1,18 @@
+import { UserId, EventId } from './index'
+
+export type SearchParams = {
+  category: string
+  location: {
+    lat: string
+    lng: string
+  }
+  distance: string
+}
+
+export type ProfilePageParams = {
+  userId: UserId
+}
+
+export type EventEditPageParams = {
+  eventId: EventId
+}
