+++ client/src/components/NavDropdownFunction/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
-import { NavDropdownFunctionProps } from '../../types'
+import { NavDropdownFunctionProps } from '../../Types'
 
 const NavDropdownFunction = ({
   text,
