+++ client/src/util/AdminRoute.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Route, Redirect } from 'react-router-dom'
-import { ProtectedRouteProps, AppState } from '../types'
+import { ProtectedRouteProps, AppState } from '../Types'
 import { useSelector } from 'react-redux'
 
 const AdminRoute = ({ component: Component, ...rest }: ProtectedRouteProps) => {
